+++ pkg/validator/controllers/statefulsets.go
@@ -2,42 +2,20 @@ package controllers
 
 import (
 	"github.com/fairwindsops/polaris/pkg/config"
+	"github.com/sirupsen/logrus"
 	kubeAPIAppsV1 "k8s.io/api/apps/v1"
-	kubeAPICoreV1 "k8s.io/api/core/v1"
-	kubeAPIMetaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-// StatefulSetController is an implementation of controller for deployments
-type StatefulSetController struct {
-	GenericController
-	K8SResource kubeAPIAppsV1.StatefulSet
-}
-
-// GetPodTemplate returns the kubernetes template spec
-func (s StatefulSetController) GetPodTemplate() *kubeAPICoreV1.PodTemplateSpec {
-	return &s.K8SResource.Spec.Template
-}
-
-// GetPodSpec returns the podspec from the original kubernetes resource
-func (s StatefulSetController) GetPodSpec() *kubeAPICoreV1.PodSpec {
-	return &s.K8SResource.Spec.Template.Spec
-}
-
-// GetObjectMeta returns the metadata
-func (s StatefulSetController) GetObjectMeta() kubeAPIMetaV1.ObjectMeta {
-	return s.K8SResource.ObjectMeta
-}
-
-// GetKind returns the supportedcontroller enum type
-func (s StatefulSetController) GetKind() config.SupportedController {
-	return config.StatefulSets
-}
-
 // NewStatefulSetController builds a statefulset controller
-func NewStatefulSetController(originalResource kubeAPIAppsV1.StatefulSet) Interface {
-	controller := StatefulSetController{}
+func NewStatefulSetController(originalResource kubeAPIAppsV1.StatefulSet) GenericController {
+	controller := GenericController{}
 	controller.Name = originalResource.Name
 	controller.Namespace = originalResource.Namespace
-	controller.K8SResource = originalResource
+	controller.PodSpec = originalResource.Spec.Template.Spec
+	controller.ObjectMeta = originalResource.ObjectMeta
+	controller.Kind = config.StatefulSets.String()
+	if controller.Name == "" {
+		logrus.Warn("Name is missing from controller", originalResource.Namespace)
+	}
 	return controller
 }