+++ pkg/validator/fullaudit_test.go
@@ -10,10 +10,12 @@ import (
 )
 
 func TestGetTemplateData(t *testing.T) {
-	k8s := test.SetupTestAPI()
+	k8s, dynamicClient := test.SetupTestAPI()
 	k8s = test.SetupAddControllers(k8s, "test")
 	k8s = test.SetupAddExtraControllerVersions(k8s, "test-extra")
-	resources, err := kube.CreateResourceProviderFromAPI(k8s, "test")
+	// TODO figure out how to mock out dynamic client.
+	// and add in pods for all controllers to fill out tests.
+	resources, err := kube.CreateResourceProviderFromAPI(k8s, "test", &dynamicClient)
 	assert.Equal(t, err, nil, "error should be nil")
 
 	c := conf.Configuration{
@@ -33,8 +35,8 @@ func TestGetTemplateData(t *testing.T) {
 
 	sum := CountSummary{
 		Successes: uint(0),
-		Warnings:  uint(9),
-		Errors:    uint(9),
+		Warnings:  uint(1),
+		Errors:    uint(1),
 	}
 
 	actualAudit, err := RunAudit(c, resources)
@@ -48,17 +50,7 @@ func TestGetTemplateData(t *testing.T) {
 		kind    string
 		results int
 	}{
-		{kind: "Deployment", results: 2},
-		{kind: "Deployment", results: 2},
-		{kind: "Deployment", results: 2},
-		{kind: "StatefulSet", results: 2},
-		{kind: "StatefulSet", results: 2},
-		{kind: "StatefulSet", results: 2},
-		{kind: "DaemonSet", results: 2},
-		{kind: "DaemonSet", results: 2},
-		{kind: "Job", results: 0},
-		{kind: "CronJob", results: 0},
-		{kind: "ReplicationController", results: 2},
+		{kind: "Pod", results: 2},
 	}
 
 	assert.Equal(t, len(expected), len(actualAudit.Results))